What I did
I added a border around the playable area. 

Later, I'll need to expand the window size so that I can add the counts of good and bad food above the game. 

I had to scrap my initial plan on how i was gonna make the counters because it broke the game lol. So I'm just gonna add this for now